Auto style now prioritizes example sections.
This allows a docstring to contain examples that contain
docstrings of other styles. Previously this would cause the whole
docstring to be detected as that style. Now the number one
criteria is which style has more examples. If there is a tie
then the original criterium of most meta items is fallen back to.